Split up EVY swift
