+++ ios/evy/Core/EVY+TextParsing.swift
@@ -1,0 +1,43 @@
+//
+//  EVY+TextParsing.swift
+//  evy
+//
+
+import Foundation
+
+extension EVY {
+  static func getDataFromText(_ input: String) throws -> EVYJson {
+    try _getDataFromText(input)
+  }
+
+  static func getDataFromProps(_ props: String) throws -> EVYJson {
+    try _getDataFromProps(props)
+  }
+
+  static func getValueFromText(_ input: String, editing: Bool = false) throws -> EVYValue {
+    try _getValueFromText(input, editing: editing)
+  }
+
+  static func parsePropsFromText(_ input: String) -> String {
+    _parsePropsFromText(input)
+  }
+
+  static func watchTarget(for text: String) -> String {
+    _watchTarget(for: text)
+  }
+
+  static func evaluateFromText(_ input: String) throws -> Bool {
+    try _evaluateFromText(input)
+  }
+
+  static func formatData(json: EVYJson, format: String) throws -> String {
+    try _formatData(json: json, format: format)
+  }
+
+  static func formatDataOrToString(json: EVYJson, format: String) throws -> String {
+    if format.isEmpty {
+      return json.toString()
+    }
+    return try formatData(json: json, format: format)
+  }
+}
